+++ shifter/shifter_platform/ctf/migrations/0006_rename_ctf_flag_challen_d1c5e7_idx_ctf_flag_challen_f6fd67_idx_and_more.py
@@ -7,7 +7,7 @@ from django.db import migrations, models
 class Migration(migrations.Migration):
 
     dependencies = [
-        ('ctf', '0003_ctf_flag_model'),
+        ('ctf', '0005_ctf_programmable_flag_validation'),
     ]
 
     operations = [